test(dropdown): cover overlay handoff in the existing jsdom suite
