+++ packages/react/tests/components/dropdown/dropdown.test.tsx
@@ -1,6 +1,11 @@
 import {User, cleanup, render, runAllTimers, screen} from "@heroui/testing/helpers";
 
-import {DropdownFixture} from "./fixtures";
+import {
+  DropdownFixture,
+  DropdownToAlertDialogFixture,
+  DropdownToDrawerFixture,
+  DropdownToModalFixture,
+} from "./fixtures";
 
 describe("Dropdown", () => {
   let testUtilUser: User;
@@ -93,5 +98,65 @@ describe("Dropdown", () => {
 
     expect(tester.getMenu()).toBeNull();
     expect(onAction).not.toHaveBeenCalled();
+  });
+
+  /** Regressions for the leftover dropdown overlay reported in #6344 and #6361. */
+  describe("handing off to a modal-level overlay", () => {
+    const selectItem = async (option: string) => {
+      const tester = testUtilUser.createTester("Menu", {
+        root: screen.getByRole("button", {name: "Menu"}),
+      });
+
+      await tester.open();
+      runAllTimers();
+
+      await tester.toggleOptionSelection({option});
+      runAllTimers();
+
+      return tester;
+    };
+
+    it("supports opening a Modal from an item without leaving the popover behind", async () => {
+      render(<DropdownToModalFixture />);
+
+      const tester = await selectItem("Open modal");
+
+      expect(tester.getMenu()).toBeNull();
+      expect(document.querySelector('[data-slot="dropdown-popover"]')).toBeNull();
+      expect(screen.getByRole("dialog", {name: "Modal from dropdown"})).toBeInTheDocument();
+    });
+
+    it("supports closing a controlled dropdown while opening an AlertDialog", async () => {
+      render(<DropdownToAlertDialogFixture />);
+
+      const tester = await selectItem("Delete project");
+
+      expect(tester.getMenu()).toBeNull();
+      expect(document.querySelector('[data-slot="dropdown-popover"]')).toBeNull();
+      expect(screen.getByRole("alertdialog")).toBeInTheDocument();
+    });
+
+    it("supports opening a Drawer from an item without leaving the popover behind", async () => {
+      render(<DropdownToDrawerFixture />);
+
+      const tester = await selectItem("Open drawer");
+
+      expect(tester.getMenu()).toBeNull();
+      expect(document.querySelector('[data-slot="dropdown-popover"]')).toBeNull();
+      expect(screen.getByRole("dialog", {name: "Drawer from dropdown"})).toBeInTheDocument();
+    });
+
+    it("keeps a controlled dropdown closed once the AlertDialog holds focus", async () => {
+      render(<DropdownToAlertDialogFixture />);
+
+      const tester = await selectItem("Delete project");
+
+      // React Aria restores focus to the trigger when the popover unmounts; that must not
+      // resurrect the controlled dropdown behind the dialog.
+      runAllTimers();
+
+      expect(tester.getMenu()).toBeNull();
+      expect(screen.getByRole("alertdialog")).toBeInTheDocument();
+    });
   });
 });
